Factor out util.py
