+++ bazel/verilog_lint_test.py
@@ -13,19 +13,8 @@
 # limitations under the License.
 
 import argparse
+from util import check_each_filename_suffix
 from typing import List, Optional
-
-def check_each_filename_suffix(filenames: List[str], suffices: List[str]) -> None:
-    """Raises ValueError if there is any filename in the list that doesn't end with any of the expected suffices."""
-    for filename in filenames:
-        match = False
-        for suffix in suffices:
-            if filename.endswith(suffix):
-                match = True
-                break
-        if not match:
-            raise ValueError(f"Expected filename to end with any of {suffices}: {filename}")
-
 
 def verilog_lint_test(hdrs: Optional[List[str]], srcs: List[str]) -> bool:
     """Returns True if the the Verilog/SystemVerilog sources pass a lint test; may print to stdout and/or stderr."""
